7.20190319   BT_LOST_CC2540F256_V07_20190319(LOST-BLE) a.修改电量检测为ad,按键改为以前的

+++ 调试案子/BT-LOST/Firmware/BT-LOST-CC254X/Projects/ble/common/cc2540/OnBoard.h
@@ -75,7 +75,7 @@ typedef struct
 {
   osal_event_hdr_t hdr;
   uint8             state; // shift
-  uint16             keys;  // keys  //ABC
+  uint8             keys;  // keys
 } keyChange_t;
 
 // Timer clock and power-saving definitions
